(refactor): move tempaltes to seperate files for better maintainability
